reorganize/de-cruft tests
